Updating getting support on the forums
Old documentation had no current link to the Counterparty Forums and said "Support" area was for Private Support when it is in fact public. 

Also updated the screenshots to the new UI of the forums, as well as updated the Private Support to be messaging the current Admin and Moderator of the forums.